requirements.txt
img2table==1.4.1
requests==2.32.3
langchain==0.3.21
langchain-core==0.3.48
fastapi==0.115.0
python-multipart==0.0.9
pydantic==2.10.6
anyio==4.9.0
httpx==0.28.1
uvicorn==0.34.0
openai==1.68.2
langchain_openai
langchain_community
pillow==11.2.0
pytesseract==0.3.13

pymupdf==1.24.10
qreader==3.1
opencv-python==4.10.0.84
python-dotenv==1.0.1
pyzbar==0.1.9